webauthn: add interface for managing local credentials.
A settings UI for managing local credentials will need a wrapper that
abstracts over platform-specific differences. This change adds such a
wrapper. At the moment only Windows is supported, so the
platform-agnostic bit is merely "potential" rather than "actual". But it
includes code to handle the fact that we'll need to quickly know if
credentials exist (in order to decide whether to show a link on the
passwords page) but enumerating credentials on Windows gets rather slow
if there are too many of them.

Change-Id: Icddc3170f962a70aae9fc432ddd3d67e5289bdda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3753438
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1023971}