Use Coordinator pattern for ExtensionsTabbedMenuCoordinator
This CL should not change any behavior in ExtensionsTabbedMenuView.

Architectural changes:
- ExtensionsTabbedMenuView's lifetime can be affected by both the
ExtensionsToolbarContainer as well as two different
ExtensionsToolbarButtons. We extract an
ExtensionsTabbedMenuCoordinator interface so that all showing and
hiding of the ExtensionsTabbedMenuView (outside of things like
deactivation) are controlled by the coordinator.

- ExtensionsToolbarContainer owns the coordinator since it owns the
ExtensionsToolbarButtons as well.

- We eliminate the namespace raw pointer to ExtensionsTabbedMenuView
and the static methods on ExtensionsTabbedMenuView. The state
of the ExtensionsTabbedMenuView is stored within the class itself
rather than the namespace raw pointer.

- We do not need to check IsShowing()
within ExtensionsTabbedMenuView because the object will only be
constructed when it will be shown.

Bug: 1342246
Change-Id: Ie1b151d27bad6fb56b9c82438657bfb2fd1cc1da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3750737
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Elaine Chien <elainechien@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1023992}