HLS: Implement the EXT-X-PART tag
This CL implements the EXT-X-PART tag parsing function, along with unit
tests. Integration of this tag with the hls::MediaPlaylist class is more
complex than most other tags, so that is left to a follow-up CL. This CL
adds unit tests covering the new parsing function for this tag.

Bug: 1266991
Change-Id: I1b855295a3a8425cdc155eab1e445960d7909ccc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3759206
Commit-Queue: Will Cassella <cassew@chromium.org>
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024029}