Set test notifications only for NetworkChangeNotifier in NetworkService
This CL lets SetTestNotificationsOnly() to enable/disable non-test
notifications from NetworkChangeNotifier not only in browser process but
also in NetworkService process.

NetworkChangeNotifier in NetworkService process notifies browser process
by NetworkConnectionTracker and this was not disabled by
NetworkChangeNotifier::SetTestNotificationsOnly() which leads some test
flakiness.

NetworkChangeNotifier::test_notifications_only_ must be set before
NetworkChangeNotifier is created as commented, so it is correct to
refer to it on NetworkChangeManagerClient construction.

Bug: 1170609
Change-Id: I33b7ede5ce4f52abdfae8b542edbf25deaaacce4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3758296
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024113}