HighlightOverlayPainting: optimise spelling-and-grammar-only paints
Use cases involving many spelling errors are very common, and they are
especially important for input latency. For example, the user might be
typing in a textarea with the dictionary set to the wrong language.

This kind of scenario regresses under HighlightOverlayPainting, since
the old impl (incorrectly) just slaps a squiggly line onto the text
without breaking the text into parts and painting in overlay order.

The good news is, if a text fragment has no highlights that change the
text color, add backgrounds, or add decorations that paint under some
decorations from earlier layers [1], then we can do exactly that: it
is equally correct to paint the whole text fragment then slap some
decorations onto the text.

This patch bypasses HighlightOverlayPainting when the only highlights
are spelling and/or grammar errors, and they are completely unstyled.
The default style only adds spelling/grammar error decorations, which
css-text-decor-4 doesn’t prescribe any particular painting order for,
and which we already paint on top of other web-exposed decorations.

Input latency numbers are as follows, using the same methodology as in
<https://crbug.com/1335762#c21> where v3 is with CL:3731467 only and
v4 is with this patch as well:

• off: mean 12950 13489 13333*, percent_95 23758* 24522 23064
• v3: mean 14261 14849 14447*, percent_95 24498 25019* 25586
• v4: mean 14235 13548* 13181, percent_95 25492 24468* 23147
• v3 vs off (*): mean +8.36%, percent_95 +5.31%
• v4 vs off (*): mean +1.61%, percent_95 +2.99%

[1] e.g. ::target-text is later than originating, but its underlines
must paint under originating overlines (csswg-drafts#6022)

Bug: 1147859
Fixed: 1335762
Change-Id: I26ab7b32a00f0f7091cb3dd68cac2e36c997da42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3759693
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Delan Azabani <dazabani@igalia.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024120}