Set arm version 8 as default lacros ARM version.
Currently, lacros ARM is built with arm 7. It turns out all chromeos
supported ARM devices is arm 8 (or later), and chrome built with
arm 8 is faster.

This is not yet in effect practically. We have to remove the overriding
config from *-crostoolchain.gni, which will be done in another CL.

BUG=1325378
TEST=Built locally.

Change-Id: I17c6875005f3c5ae1038022468c78b316ecd69f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3759798
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024130}