Remove unused banner strings and unnecessary TODOs
The holding space banner had strings that were being used in tablet
mode, however, changes in the banner framework caused them to no longer
work. They were slightly modified and a variant created, now that the
banner framework is on by default we don't require the old variants.

Another TODO was left to update the list of allowed volume types however
this list is used elsewhere in the code so it is best to leave it as is.
To avoid the TODO getting stale, remove it.

Bug: None
Change-Id: I4d05ac9622c37b78b2723ab2e3a5c72311892368
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3760057
Auto-Submit: Ben Reich <benreich@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024136}