HDR Metadata: Plumb metadata using CALayerOverlay
When drawing a software decoded HDR video frame using
AVSampleBufferDisplayLayer, we need to pass HDR metadata in order for
CoreAnimation to render it correctly.

This was originally implemented by attaching HDR metadata to the
IOSurface in the renderer process, and then pulling this metadata out
from the IOSurface in the GPU process, effectively using the IOSurface
itself as a IPC side-channel.

Now that HDR metadata is appropriately set on YUVVideoDrawQuads and
TextureDrawQuads, propagate that metadata through viz::CALayerOverlay,
to ui::CARendererLayerParams, to ui::CARendererLayerTree:ContentLayer,
and finally to the AVSampleBufferDisplayLayer.

Leave the old IOSurface-based path in place. It will be removed in
a follow-on patch.

Bug: 1325384
Change-Id: Iafb918d078ac17e29c9eedd6caef03d9f38dab30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3757682
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024175}