HDR Metadata: Remove IOSurface IPC side channel
Historically, HDR Metadata was propagated from the
GpuMemoryBufferVideoFramePool to the CARendererLayerTree by attaching
it to the IOSurface, effectively creating an IPC side-channel between
the renderer and GPU process.

We now propagate this data through the same channel that the rest
of the CARendererLayerTree's parameters are propagated. Delete the
old path.

Bug: 1325384
Change-Id: If3b083cb5e7de4d1efb2a1c3cbb229cd8c843b8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3757717
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024186}