[Identifiability Study] Factor common logic in browser tests.
The Identifiability Study (a.k.a. Privacy Budget project) has several
browser tests, which rely upon the UKM infrastructure at different
levels for testing the study's data collection logic. In particular:

1) Some tests need to check that UKM metrics are sampled in the
client. These tests rely on a TestUKMRecorder, which adds itself as a
delegate to the DelegatingUKMRecorder.

2) Some other tests need to check the actual content of the final UKM
reports. These tests need to enable the actual UKM recording and
inspect the final report. From some point of view, this is more
complex than the former.

Previously we had one base class for the tests which was doing a
combination of 1) and 2). This CL clearly splits the two things into
separate test classes. Moreover, it removes duplicated logic in the
browser tests for the REID score estimation, making them rely on the
same base class as in 1).

Change-Id: I90deb7fb77b2aa4768d1e55bc0be6925c2063615
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3757763
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024214}