Propagate prefers-color-scheme override to popup page
Currently, “Enable automatic dark mode” feature doesn’t work correctly
for select popups because when that feature is enabled
instead of changing the page’s preferred color scheme
setting, we’re setting a media feature override[1]

However, this information is not propagated to the newly
created page for the popup.

It works correctly in other popups because they use
`owner style` to decide on the preferred color scheme[2] which
takes the media override into account[3]

This CL propagates the media override to the popup page
which in turn shows a dark mode view.

Though there is one different thing between the
behavior of the “Enable automatic dark mode” and a page
setting `color-scheme` meta to `dark” (DevTools dark[4], meta dark[5])
which I suspect the issue is, we’re setting forced dark mode colors as well
and in turn, it disables updating `backgroundColor` and other color information
here[6], so I wonder if we should do something about it.

[1]: https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/inspector/inspector_emulation_agent.cc;l=331?q=inspector_emulation_agent.cc&ss=chromium
[2]: https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/page/page_popup_client.cc;l=210?q=file:cc$%20page_popu&ss=chromium
[3]: https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/css/style_engine.cc;l=3092;bpv=0;bpt=1?q=style_engine.cc&ss=chromium%2Fchromium%2Fsrc
[4]: https://bugs.chromium.org/p/chromium/issues/attachment?aid=560959&signed_aid=WK8ZsI56vwduFSCn8Lbn8w==&inline=1&thumb=1
[5]: https://bugs.chromium.org/p/chromium/issues/attachment?aid=560958&signed_aid=ezFIO-WJonAlCXS8EWP_Mg==&inline=1&thumb=1
[6]: https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/html/forms/internal_popup_menu.cc;l=187?q=file:cc$%20internal_popup_menu&ss=chromium

Fixed: 1311561
Change-Id: Ic7ec60a0d9bb6a9b89e74ffea2ff864b3f99ee60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3745525
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024228}