Take into account caching when calculating request/body sizes
The code currently assumes that URLResponseInfo#getReceivedByteCount
returns the response size even for cached responses. This is not the
case instead, if the response is served from cached it will return 0.
This leads to a bug where we report negative response body size.

Fix this by checking if the response has been served from cache, in
which case we explicitly report response size 0.

Bug: b:226554121
Change-Id: Id591dc5b487de0903261d79f723baaddc7d50871
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3757753
Commit-Queue: Stefano Duo <stefanoduo@google.com>
Reviewed-by: Richard Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024267}