[Multipaste] Collect stack trace from the failed flaky tests
Right now, `ClipboardHistoryWebContentsBrowserTest.
VerifyHTMLRendering` is flaky and it cannot be reproduced on my
local workstation even with the CQ's GN args.

In order to collect the stack trace from the CQ running, this CL
does the following things:

(1) Enable the flaky test

(2) Add a callback that runs when the clipboard history menu is closed
then leads to a crash. The callback's life cycle is carefully
controlled so that the callback runs only when the original test
fails.

This CL will be reverted soon when collecting enough data.

Bug: 1224777
Change-Id: I0e2d8ae5c7ec9d68fa37e6cd4cae392655f166dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3761273
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024304}