[Launcher Sort] Handle null index range in reorder animation
Previously, the launcher sort animation assumes that there must be
apps within the visible range. However, this may not be true as apps
could be removed during the sort animation. This cl handles that edge
case by creating an empty animation builder and notifying the callback
that the fade-in animation is aborted.
A test is added to test this special case.

Bug: 1289411
Change-Id: Ibf886a083bfd6ab94f827312430c3b7bac32bed4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3756581
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Wen-Chien Wang <wcwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024328}