[Fast Pair] Log result of updating opt-in status after pairing
A feature of the Saved Devices page is to update the user's opt-in
status to opted-in whenever a user pairs a device via Fast Pair. This
CL adds a metric to log the success or failure of updating the opt in
status for each pairing protocol.

Bug: b/236044463
Test: Verified on chrome://histograms
Change-Id: Ibe8eeedb1f1bea3fdd9d63a34854b52594d0f312
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3751587
Reviewed-by: Daniel Classon <dclasson@google.com>
Commit-Queue: Juliet Lévesque <julietlevesque@google.com>
Reviewed-by: Crisrael Lucero <crisrael@google.com>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024391}