[ntp] Remove NtpModules feature flag
Previously, NtpModules and a module-specific feature flag had to be on
for a module to be enabled. This has caused confusion and complicates
Finch studies. With this change, turning on a module-specific feature
flag is sufficient to enable the module. Any global module UI will be
automatically enabled as soon as any module module is enabled.

Bug: 1344359
Change-Id: I1256f9be2b517b34e9819c99644dd0f7beb55614
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3763249
Commit-Queue: Tibor Goldschwendt <tiborg@chromium.org>
Auto-Submit: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024417}