rgbkbd: fix for backlight flash when switching users
This is an edge case where one of the active users selects the
wallpaper extracted color option, and tries toggling between this
session and another active session. |OnActiveUserPrefServiceChanged| is
invoked but the wallpaper color may not be present at this state.

Add a fix which skips setting the backlight with wallpaper color in
|OnActiveUserPrefServiceChanged|, and lets |OnWallpaperColorsChanged|
handle setting wallpaper color when it is available.

BUG=b/238772994
TEST=ash_unittests "KeyboardBacklightColorControllerTest.*"

Cq-Include-Trybots: luci.chrome.try:linux-chromeos-chrome
Change-Id: I0282510ef7dd8a23f2ea468c9dbb3ade12af2fd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3758891
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Commit-Queue: Jason Thai <jasontt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024478}