[CrOS SLP] Remove TODO for notification opening cellular details view.
The properties of a PIN locked, and more specifically a PUK locked SIM
may not be ready in time to be shown to the user when the notification
is clicked. This causes settings to route to the home OS Settings page
immediately instead of showing the details view at [1]. This race condition happens infrequently, but does occur when switching between eSIM profiles that are locked and unlocked on the same EUICC.

As such, the SLP notification will open the Mobile data page with the sim lock dialog surfaced following the existing behavior. Furthermore, this follows the existing pattern for when users click the unlock
button in quick settings or in the cellular mobile data page.

Note that [1] will be abandoned.

[1] http://crrev/c/3761159

Bug: b/228093904
Change-Id: I6171754aef40a32d8895bd315c3d8b35bc0bc0f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3763437
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024513}