[omnibox] Track & limit trie growth by maintaining & capping terminals
This CL prevents the possibility of unbounded collection growth by
enforcing hard limits on how many domain names (marked by terminals
in trie structure) can be inserted. The limits are based on existing
work around OmniboxFieldTrial::MaxNumHQPUrlsIndexedAtStartup and
that method is used as a secondary cap in case it's lower than the
constants used by the fuzzy provider. Usually it's 20000, much higher,
but mobile may have different constraints.

To support limits at all requires implementation of terminal counting
in the trie operations, so that is a main focus of this change. Some
code and comment cleanup is also included as this model appears to be
working better than expected and we plan to attempt launch of the
in-memory fuzzy matching approach.

Bug: 1286859
Change-Id: Ied1b51eec9b312c6a02452869bd28193416cce40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3763605
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024554}