Extract: Enable extraction cancellation
Adds a task runner to UnzipperImpl to allow mojo disconnection messages
to get through while ZIP extraction is in progress.

Adds a wrapper class ZipFileUnpacker on the client side of the unzipper
service so that the client can manage the lifetime of extraction tasks.

Refactor of the zip unpacker code to allow direct calls as before
(unzip::Unzip) or to use the wrapper class.

Lifetime of the legacy UnzipParams struct is maintained on the new
task runner and ZipFileUnpacker provides API to cancel ZIP extraction
operations that are in progress.

Bug: 953256
Tests: Manually tested extensively, including Chrome extension install
Change-Id: Ia456059a867abf1a654e327e4e0cf11618c07337
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3750826
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024568}