[Color Pipeline] Fix delay in WebContents getting correct ColorProviderSource
This CL sets the ColorProviderSource on the WebContents when it is
inserted into the TabStripModel. At this point we know which
BrowserView the WebContents will be associated with. If we move
ColorProvider code out of views/ then creation of the
ColorProviderBrowserHelper should also be moved out of views/
(potentially to the Browser class).

The reason this CL is necessary is because currently we only set the
ColorProviderSource when the WebContents is associated with the
WebView. However, this only happens when the user navigates to the
tab that will display the WebContents. This does not always happen
immediately when the tab is created. For example when the tab is
opened up through a bookmarks folder. (See bug). With this CL, the
WebContents will have the correct ColorProviderSource before the user
navigates to the tab.

Bug: 1341905
Change-Id: Ia9749833d166546b3a4e8840964d1f8eba52c060
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3751977
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Commit-Queue: Elaine Chien <elainechien@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024591}