[reland][omnibox] Capture escape keys when popup is open.
Relands crrev.com/c/3761156 .
Which was reverted with crrev.com/c/3762407 due to
linux-chromeos-chrome bot failure.
The bot failure is fixed by explicitly enabling a feature in the failing
test. Maybe, field_trial_config.json wasn't sufficient for some reason?

Currently, the omnibox captures the escape key only when input is in
progress, but not when the popup is open without input.

(The popup being open also means the omnibox has focus. If the omnibox
loses focus, the popup is closed.)

Therefore, when both the omnibox popup and find dialog are open, without
omnibox input, escape closes the find dialog rather than the omnibox
popup. A 2nd escape then closes the popup.

This CL changes the omnibox to capture escape when either the input is
in progress OR (it is focused AND the kBlurWithEscape feature is
enabled).

This fix is limited to when the kBlurWithEscape feature is enabled,
because, otherwise, the omnibox never loses focus and indefinitely
captures escape presses. That may still be better than current behavior,
but it's less clear.

This has multiple benefits:

1) It makes the escape handling consistent between zero-suggest and
   non-0-suggest.
   - Before, with 0-suggest and an omnibox selection, 1st the selection
     would be cleared, then find closed, then the omnibox closed.
   - Before, with non-0-suggest and an omnibox selection, 1st the
     selection would be cleared, then the omnibox closed, then find
     closed.
   - With this CL, in both 0 and non-0, 1st the selection is closed,
     then the omnibox closed, then the omnibox blurred, then find
     closed.

2) It makes omnibox-v-find order consistent.
   - Before, sometimes find was closed before the omnibox (e.g. with no
     user input).
   - Before, sometimes find was closed after the omnibox (e.g. with user
     input).
   - Before, sometimes find was interleaved between omnibox actions
     (e.g. with an omnibox selection).
   - With this CL, find is always last.

Exhaustively, the following documents ordering for all combinations of:
before/after this CL; with kBlurWithEscape enabled/disabled; with 0 and
non-0-suggest; and with neither, either, or both omnibox input and
selection:

Before CL
  Feature disabled
    0-suggest
      no input = find, popup
      input = input + popup, find
      selection = selection, find, popup
      input + selection = selection, input + popup, find
    non-0-suggest
      no input = find, elide
      input = input + popup, find
      selection = selection, popup, find
      input+selection = selection, input + popup, find
  kBlurWithEscape enabled
    0-suggest
      no input = find, popup, blur
      input = input + popup, find, blur
      selection = selection, find, popup, blur
      input + selection = selection, input + popup, find, blur
    non-0-suggest
      no input = find, blur
      input = input + popup, find, blur
      selection = selection, popup, find, blur
      input+selection = selection, input + popup, find, blur
With CL
  kBlurWithEscape disabled
    0-suggest
      no input = find, popup
      input = input + popup, find
      selection = selection, find, popup
      input + selection = selection, input + popup, find
    non-0-suggest (same as before patch)
      no input = find, elide
      input = input + popup, find
      selection = selection, popup, find
      input+selection = selection, input + popup, find
  kBlurWithEscape enabled
    0-suggest
      no input = popup, blur, find
      input = input + popup, blur, find
      selection = selection, popup, blur, find
      input + selection = selection, input + popup, blur, find
    non-0-suggest
      no input = popup, blur, find
      input = input + popup, blur, find
      selection = selection, popup, blur, find
      input + selection = selection, input + popup, blur, find

Bug: 776080, 935157
Change-Id: I0d4418bedf77b78757a497a84ef57dce17102665
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3763500
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024614}